Migrate conversion of string to array to utility function [skip ci]
