+++ src/utils/index.ts
@@ -1,2 +1,3 @@
+export { default as convertMvStringToArray } from './convertMvStringToArray';
 export { default as dummyLogger } from './dummyLogger';
 export { default as ensureArray } from './ensureArray';
